tests, network: Extract SR-IOV tests to a dedicated file
