+++ tests/network/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
         "primary_pod_network.go",
         "probes.go",
         "services.go",
+        "sriov.go",
         "vmi_infosource.go",
         "vmi_istio.go",
         "vmi_lifecycle.go",
